Fixing #26 - Support for ordered joinsets
Even tho a "set" is inherantly an unordered collection of files,
it can be desired to join files together in a specific order to
allow proper Javascript evaluation when compiled into a single
function/closure.

This patch adds a new <orderedFiles/> list of <file/> entries to a
<joinset/> in additional to the existing fileset.

In order for this to work properly however, <compileIndividualFiles/>
must be false ( the default value ).

Change-Id: I63c36a611b9f38a48a6e298469a74ee67fedf97e